Refactor

+++ Steve/GameEngine/External/lib/External/include/Common/Typedef.h
@@ -3,8 +3,8 @@
 using uint8  = unsigned char;
 using uint16 = unsigned short;
 using uint32 = unsigned int;
-using uint64 = long unsigned int;
-using int8   = char;
-using int16  = short;
-using int32  = int;
-using int64  = long int;
+using uint64 = unsigned long long;
+using int8   = signed char;
+using int16  = signed short;
+using int32  = signed int;
+using int64  = signed long long;
